Adds "structured concurrency" on generated sample
Is not exactly structured concurrency, but explores how to
make a boc wait until all the internal bocs are completed while
executing then concurrenly.

It stills need more work as every inner boc will have to do the same but
proves that is possible to run each boc concurrently and still wait
until the whole block finishes.

I need to figure out how to wait if a value is expected

`a:func()` vs `func()`