Arrays needs comma between elements.
Because of the automatic comma insertion, array now need comma after each element.
A possible solution for this would be to remove the automatic comma insertion which we might need to revisit in the future.
For the time being, adding it would simplify the parsing and would save some capital of the "strange budget"